eps 11

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,10 @@ use App\Post;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -73,6 +77,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page.');
+        }
         return view('posts/edit')->with('post', $post);
 
     }
@@ -108,6 +115,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page.');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'post removed');
     }
